% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.create_data}
\alias{.create_data}
\title{Wrapper for data creation}
\usage{
.create_data(covar_true, N)
}
\arguments{
\item{covar_true}{covariance matrix (must be symmetric and positive definite)}

\item{N}{number of replications}
}
\value{
an N * n matrix with each row corresponding to a replication of all covariates/locations
}
\description{
Improves readability for data creation rather than having to call mvrnorm each time. Also, 
there are many alternatives to mvrnorm in R, which isn't important for the package usefulness.
Lastly, I was curious to see how internal functions work, as it is "good coding practice."
}
\examples{

create_data(diag(2), 10)

}
\keyword{internal}
