% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderfuns.R
\name{find_nn}
\alias{find_nn}
\title{Find nearest neighbors based on a tapered covariance matrix}
\usage{
find_nn(locs, datum, m, tapering_range = 0.5)
}
\arguments{
\item{locs}{matrix of locations of points (to match input argument of fields::rdist)}

\item{datum}{Data where column i corresponds to observations at the i'th point of locs}

\item{m}{number of neighbors}

\item{tapering_range}{Percentage of the maximum distance for Exponential tapering, which
defaults to 1/2 the maximum distance.}
}
\value{
a matrix of nearest neighbors of dimension n x m
}
\description{
Find nearest neighbors based on a tapered covariance matrix
}
