% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderfuns.R
\name{find_nn_dist}
\alias{find_nn_dist}
\title{Find nearest neighbors based on a distance matrix}
\usage{
find_nn_dist(d, m)
}
\arguments{
\item{d}{distance matrix (e.g. distances)}

\item{m}{number of neighbors}
}
\value{
a matrix of nearest neighbors of dimension n x m
}
\description{
Find nearest neighbors based on a distance matrix
}
