% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_map}
\alias{get_map}
\title{Creates MAP (maximum a posteriori) sparse matrix from thetas}
\usage{
get_map(thetas, datum, NNarray, threshh = 0.001)
}
\arguments{
\item{thetas}{3 real numbers representing the three hyperparameters}

\item{datum}{an N * n matrix of the data (N replications of n locations/variables)}

\item{NNarray}{an n * m2 matrix giving the m nearest neighbors previous in the ordering (or
outputting NAs if not available [i.e. there are not m previous points]) that are ordered
from closest to furthest away. It is OK to have m2 large, as it will be reduced to match the size
of the posterior means (i.e. number of columns in the third element of the posteriors), but
never have m2 < 2.}

\item{threshh}{threshold for number of neighbors (for thetas_to_priors); defaults
to 1e-3}
}
\description{
This is basically a combination of thetas_to_priors, get_posts, and samp_posts so one does not
have to store the posteriors in memory (that is more for Bayesian uses).
}
