% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{na_omit_c}
\alias{na_omit_c}
\title{Remove NAs from a vector}
\usage{
na_omit_c(x)
}
\arguments{
\item{x}{a vector (possibly with NAs)}
}
\description{
This function is needed for the direct translation of the R functions to C++. This function is 
taken from StackOverflow approximately: https://stackoverflow.com/questions/19156353/remove-na-values-efficiently
}
\examples{
na_omit_c(c(NA,5,10,NA,303))
}
