\name{NPVecchia-package}
\alias{NPVecchia-package}
\alias{NPVecchia}
\docType{package}
\title{
\packageTitle{NPVecchia}
}
\description{
\packageDescription{NPVecchia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NPVecchia}
\packageIndices{NPVecchia}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NPVecchia}

Maintainer: \packageMaintainer{NPVecchia}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
