% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster_orderings.R
\name{order_maximin_dist}
\alias{order_maximin_dist}
\title{Maximin ordering based on a distance}
\usage{
order_maximin_dist(d)
}
\arguments{
\item{d}{distance matrix (e.g. distances)}
}
\value{
Maximin ordering of the points corresponding to the matrix
}
\description{
This function orders the points in a maximin ordering based on distance.
}
\examples{
n <- 100
fake_dists <- matrix(runif(n^2), nc = n)
#make it a valid distance matrix by making d(a,a) = 0
diag(fake_dists) <- 0
order_mm <- order_maximin_dist(fake_dists)
#reorder data
}
