% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster_orderings.R
\name{order_mm_covariance}
\alias{order_mm_covariance}
\title{Maximin ordering using covariance matrix}
\usage{
order_mm_covariance(cov.matrix)
}
\arguments{
\item{cov.matrix}{a covariance matrix}
}
\value{
Maximin ordering of the points corresponding to the matrix
}
\description{
This function is a wrapper for \code{\link{order_maximin_dist}} to take in a covariance
matrix as input, convert it to a distance as 1 - correlation, and thene calculates the 
ordering based on that distance. \code{\link{order_mm_tapered}} is preferred unless
one has a lot of replications as tapering the sample covariance improves performance.
}
\examples{
datum <- matrix(rnorm(1e3), nc=10)
sample_cov <- cov(datum)
order_datum <- order_mm_covariance(sample_cov)
}
