% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderfuns.R
\name{orderByCoordinate}
\alias{orderByCoordinate}
\title{Order by one a single coordinate}
\usage{
orderByCoordinate(locs, coordinate)
}
\arguments{
\item{locs}{locations to order}

\item{coordinate}{index of the coordinate to order by}
}
\description{
Order by one a single coordinate
}
\keyword{internal}
