% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderfuns.R
\name{orderDist2Point}
\alias{orderDist2Point}
\title{Order by distance to some point loc0}
\usage{
orderDist2Point(locs, loc0)
}
\arguments{
\item{locs}{locations to order by distance to point loc0}

\item{loc0}{reference point for ordering locations}
}
\description{
Order by distance to some point loc0
}
\keyword{internal}
