% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderfuns.R
\name{orderMaxMinFast}
\alias{orderMaxMinFast}
\title{Max-min ordering function}
\usage{
orderMaxMinFast(locs, numpropose)
}
\arguments{
\item{locs}{matrix of locations}

\item{numpropose}{the number of locations to order}
}
\value{
orderinds a vector to order the locations
}
\description{
This is the O(n^2) algorithm for AMMD ordering. Start with a point
in the middle, then propose a random set of the remaining points
(of size 'numpropose') and choose the one that has maximum minimum
distance to the already selected points. set 'numpropose' to n
to get the exact maximum minimum distance ordering
}
