% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderfuns.R
\name{orderMiddleOut}
\alias{orderMiddleOut}
\title{Order by distance to the center}
\usage{
orderMiddleOut(locs)
}
\arguments{
\item{locs}{locations to order}
}
\description{
Order by distance to the center
}
\keyword{internal}
