% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{samp_posts}
\alias{samp_posts}
\title{Creates posterior mean sparse matrix from posteriors}
\usage{
samp_posts(posts, NNarray)
}
\arguments{
\item{posts}{a List of the posteriors from get_posts (or get_posts_c); alternatively it can be
custom values as long as the sizes match the output from get_posts.}

\item{NNarray}{an n * m2 matrix giving the m nearest neighbors previous in the ordering (or 
outputting NAs if not available [i.e. there are not m previous points]) that are ordered 
for nearest to furthest. It is OK to have m2 large, as it will be reduced to match the size
of the posterior means (i.e. number of columns in the third element of the posteriors), but
never have m2 < 2.}
}
\value{
Sparse triangular matrix that is the Cholesky of the precision matrix \eqn{\Omega} 
such that \deqn{\Omega = U U'}
}
\description{
This function uses the posterior arguments to create a posterior mean estimate of the Cholesky of
the precision matrix. This function creates the Bayesian version of get_mle or its variants.
}
