% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderfuns.R
\name{find_nn}
\alias{find_nn}
\title{Find nearest neighbors based on a covariance matrix}
\usage{
find_nn(cov.matrix, m)
}
\arguments{
\item{cov.matrix}{an (ordered) covariance matrix to use for finding NN based on
this distance (1 - correlation)}

\item{m}{number of (ordered) nearest neighbors to calculate}
}
\value{
a matrix of nearest neighbors of dimension n x m
}
\description{
Find nearest neighbors based on a covariance matrix
}
