% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{get_mle}
\alias{get_mle}
\title{Creates the simplist frequentist version of our method}
\usage{
get_mle(dat, NNarray)
}
\arguments{
\item{dat}{an N * n matrix, where each row corresponds to N replications for that location}

\item{NNarray}{an n * m matrix giving the m nearest neighbors previous in the ordering (or 
outputting NAs if not available [i.e. there are not m previous points]) that are ordered 
for nearest to furthest. Required: m < N, m >= 2}
}
\value{
Sparse triangular matrix that is the Cholesky of the precision matrix \eqn{\Omega} 
such that \deqn{\Omega = U U'}
}
\description{
For each regression, the MLE is used (i.e. the coefficients and residual standard error obtained from lm). 
This will only work when N > m (or lm will have issues), so call the function accordingly.
}
