% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_posts_c}
\alias{get_posts_c}
\title{Gets posteriors of Bayesian methodology}
\usage{
get_posts_c(datum, a, b, g, NNarray)
}
\description{
This is the C++ version of \code{\link{get_posts}}. See there for further documentation.
}
