% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minus_loglikeli_c}
\alias{minus_loglikeli_c}
\title{Calculates the integrated log likelihood}
\usage{
minus_loglikeli_c(thetas, datum, NNarray)
}
\description{
This is the C++ version of \code{\link{minus_loglikeli}}. See there for further documentation.
}
\details{
This function is much faster than the R version, making it preferable.
}
