\name{NPVecchia-package}
\alias{NPVecchia-package}
\alias{NPVecchia}
\docType{package}
\title{
\packageTitle{NPVecchia}
}
\description{
\packageDescription{NPVecchia}
}
\details{

The main functions are listed below (and have C++ versions by adding "_c" to the name). See the vignette for further details.

minus_loglikeli gives the integrated log-likelihood as a function of 3 hyperparameters

thetas_to_priors gives a List of the priors from the hyperparameters

get_posts gets the posteriors from these priors

samp_posts gets the MAP from the posteriors to give an estimate
}
\author{
Brian Kidd

Maintainer: \packageMaintainer{NPVecchia}
}
\references{
Eventually, this will include a link to the Arxiv paper.
}
\keyword{ package }