% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster_orderings.R
\name{order_maximin_dist}
\alias{order_maximin_dist}
\title{Maximin ordering based on a distance}
\usage{
order_maximin_dist(d)
}
\arguments{
\item{d}{distance matrix (e.g. distances)}
}
\value{
Maximin ordering of the points corresponding to the matrix
}
\description{
This function orders the points in a maximin ordering based on distance.
}
