% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faster_orderings.R
\name{order_mm_tapered}
\alias{order_mm_tapered}
\title{Maximin ordering using a tapered covariance matrix}
\usage{
order_mm_tapered(locs, datum, tapering_range = 0.4)
}
\arguments{
\item{locs}{matrix of locations of points (to match input argument of fields::rdist)}

\item{datum}{Data where column i corresponds to observations at the i'th point of locs}

\item{tapering_range}{Percentage of the maximum distance for Exponential tapering, which
defaults to 1/2 the maximum distance.}
}
\value{
Maximin ordering of the points
}
\description{
This function is a wrapper for \code{\link{order_maximin_dist}} to take in data and
locations to calculate the ordering based on a tapered sample covariance matrix. It
provides an argument to control the amount of tapering.
}
