% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{samp_posts}
\alias{samp_posts}
\title{Creates posterior mean sparse matrix from posteriors}
\usage{
samp_posts(posts, NNarray, bayesian = FALSE, uhat = TRUE)
}
\arguments{
\item{posts}{a List of the posteriors from get_posts (or get_posts_c); alternatively it can be
custom values as long as the sizes match the output from get_posts.}

\item{NNarray}{an n * m2 integer matrix giving the m nearest neighbors previous in the ordering (or 
outputting NAs if not available [i.e. there are not m previous points]) that are ordered 
from closest to furthest away. It is OK to have m2 large, as it will be reduced to match the size
of the posterior means (i.e. number of columns in the third element of the posteriors), but
never have m2 < 2.}

\item{bayesian}{a logical flag which defaults to FALSE. This controls if given the posteriors, it obtains
the MAP estimate (when FALSE) or samples all non-zero entries of the U matrix (when TRUE).}

\item{uhat}{a logical flag which defaults to TRUE. If FALSE and bayesian=TRUE, it will output a posterior
sample of the data rather than the U matrix. If bayesian=FALSE, it is unused.}
}
\value{
Sparse triangular matrix that is the Cholesky of the precision matrix \eqn{\Omega} 
such that \deqn{\Omega = U U'}
}
\description{
This function uses the posterior arguments to create a posterior mean estimate of the Cholesky of
the precision matrix. This function creates the Bayesian version of get_mle or its variants. See
the mathematical description for details of the methodology.
}
\examples{

#create fake data and fake neighbor matrix
datum <- matrix(rnorm(1e4), nrow = 10)
NNarray <- matrix(NA, nrow = 1e3, ncol = 100)
#can only use previous points in ordering (this is actually 
#impossible in low dimensional space like this is designed for)
for(i in 1:100){
  NNarray[(i + 1):1e3, i] <- i
}
priors <- thetas_to_priors(c(1, 1, 1), 1e3)

posteriors <- get_posts(datum, priors, NNarray)

uhat <- samp_posts(posteriors, NNarray)

}
