% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{samp_posts_c}
\alias{samp_posts_c}
\title{Creates posterior mean sparse matrix from posteriors}
\usage{
samp_posts_c(posts, NNarray)
}
\description{
This is the C++ version of \code{\link{samp_posts}}. See there for further documentation.
}
