% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{thetas_to_priors_c}
\alias{thetas_to_priors_c}
\title{Transforms hyperparameters to priors}
\usage{
thetas_to_priors_c(thetas, n, thresh = 0.001)
}
\description{
This is the C++ version of \code{\link{thetas_to_priors}}. See there for further documentation.
}
\details{
It is often about the same speed as the R version of this function due to little computational cost.
}
